tests/wpt/web-platform-tests/dom/nodes/DOMImplementation-createHTMLDocument-with-saved-implementation.html

<!DOCTYPE html>
<title>DOMImplementation.createHTMLDocument</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-domimplementation-createhtmldocument">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
// Test the document location getter is null outside of browser context
test(function() {
  var iframe = document.createElement("iframe");
  document.body.appendChild(iframe);
  var implementation = iframe.contentDocument.implementation;
  iframe.remove();
  assert_not_equals(implementation.createHTMLDocument(), null);
}, "createHTMLDocument(): from a saved and detached implementation does not return null")
</script>
